+++ src/cassini/__init__.py
@@ -6,7 +6,6 @@
 .. moduleauthor:: Vladimir Vukicevic <{{email}}>
 """
 
-
 from loguru import logger
 
 from cassini.cassini import __version__, connect_mqtt, get_printers, print_file, status, upload, watch