+++ src/cassini/cassini.py
@@ -6,7 +6,7 @@
 # License: MIT
 #
 import asyncio
-import pprint
+import contextlib
 import socket
 import time
 from importlib.metadata import PackageNotFoundError, version
@@ -17,6 +17,7 @@ import typer
 from loguru import logger
 from rich import print as rprint
 from rich.console import Console
+from rich.live import Live
 from rich.progress import (
     BarColumn,
     MofNCompleteColumn,
@@ -40,11 +41,14 @@ try:
 except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
 
+
 class PrintError(Exception):
     pass
 
+
 class UploadError(Exception):
     pass
+
 
 app = typer.Typer(
     name="cassini",
@@ -56,7 +60,8 @@ app = typer.Typer(
 
 verbosity_level = 0
 
-def version_callback(value: bool) -> None: # FBT001
+
+def version_callback(value: bool) -> None:  # FBT001
     """Prints the version of the package."""
     if value:
         rprint(f"[yellow]boardgamegeek[/] version: [bold blue]{__version__}[/]")
@@ -73,27 +78,53 @@ def verbosity(
             help="Control output verbosity. Pass this argument multiple times to increase the amount of output.",
             count=True,
         ),
-    ] = 0
+    ] = 0,
 ) -> None:
     verbosity_level = verbose  # noqa: F841
 
 
 async def create_mqtt_server():
-    mqtt = SimpleMQTTServer("0.0.0.0", 0)
+    mqtt = SimpleMQTTServer("127.0.0.1", 0)
     await mqtt.start()
     mqtt_server_task = asyncio.create_task(mqtt.serve_forever())
     return mqtt, mqtt.port, mqtt_server_task
 
 
 async def create_http_server():
-    http = SimpleHTTPServer("0.0.0.0", 0)
+    http = SimpleHTTPServer("127.0.0.1", 0)
     await http.start()
     http_server_task = asyncio.create_task(http.serve_forever())
     return http, http.port, http_server_task
 
 
-def do_status(printers: list[SaturnPrinter]):
+def do_status(printers: list[SaturnPrinter]) -> Table:
+    table = Table(
+        title="Status",
+        show_header=False,
+    )
     for p in printers:
+        p.refresh()
+        attrs = p.desc["Data"]["Attributes"]
+        status = p.desc["Data"]["Status"]
+        print_info = status["PrintInfo"]
+        file_info = status["FileTransferInfo"]
+
+        table.add_column("", style="green", justify="right")
+        table.add_column("", style="cyan", justify="left")
+
+        table.add_row("IP address", f"{p.addr[0]}")
+        table.add_row(f"{attrs['Name']}", f"{attrs['MachineName']}")
+        table.add_row("Machine Status:", f"{CurrentStatus(status['CurrentStatus']).name}")
+        table.add_row("Print Status:", f"{PrintInfoStatus(print_info['Status']).name}")
+        table.add_row("Layers:", f"{print_info['CurrentLayer']}/{print_info['TotalLayer']}")
+        table.add_row("File:", f"{print_info['Filename']}")
+        table.add_row("File Transfer Status:", f"{FileStatus(file_info['Status']).name}")
+    return table
+
+
+def live_status(printers: list[SaturnPrinter]):
+    for p in printers:
+        p.refresh()
         attrs = p.desc["Data"]["Attributes"]
         status = p.desc["Data"]["Status"]
         print_info = status["PrintInfo"]
@@ -101,31 +132,32 @@ def do_status(printers: list[SaturnPrinter]):
         table = Table(
             title="Status",
             show_header=False,
-            )
+        )
 
         table.add_column("", style="green", justify="right")
         table.add_column("", style="cyan", justify="left")
 
         table.add_row("IP address", f"{p.addr[0]}")
-        table.add_row(f"{attrs['Name']}",f"{attrs['MachineName']}")
-        table.add_row("Machine Status:",f"{CurrentStatus(status['CurrentStatus']).name}")
-        table.add_row("Print Status:",f"{PrintInfoStatus(print_info['Status']).name}")
-        table.add_row("Layers:",f"{print_info['CurrentLayer']}/{print_info['TotalLayer']}")
-        table.add_row("File:",f"{print_info['Filename']}")
-        table.add_row("File Transfer Status:",f"{FileStatus(file_info['Status']).name}")
-        console = Console()
-        console.print(table)
+        table.add_row(f"{attrs['Name']}", f"{attrs['MachineName']}")
+        table.add_row("Machine Status:", f"{CurrentStatus(status['CurrentStatus']).name}")
+        table.add_row("Print Status:", f"{PrintInfoStatus(print_info['Status']).name}")
+        table.add_row("Layers:", f"{print_info['CurrentLayer']}/{print_info['TotalLayer']}")
+        table.add_row("File:", f"{print_info['Filename']}")
+        table.add_row("File Transfer Status:", f"{FileStatus(file_info['Status']).name}")
+
+    return table
 
 
-def do_status_full(printers: list[SaturnPrinter]):
+def do_status_full(printers: list[SaturnPrinter]) -> None:
     for p in printers:
-        pprint.pprint(p.desc)
+        console = Console()
+        console.print_json(data=p.desc)
 
 
 def do_watch(
     printer_addr: SaturnPrinter,
-    interval: int=5,
-    ):
+    interval: int = 5,
+):
     printer = SaturnPrinter().find_printer(addr=printer_addr)
     status = printer.status()
     previous_layer = 0
@@ -144,9 +176,9 @@ def do_watch(
             # printers = SaturnPrinter.find_printers(broadcast=broadcast)
             # if len(printers) > 1:
             printer = SaturnPrinter().find_printer(addr=printer_addr)
-            status = printer.status() # I guess we really need an `update_status` method
+            status = printer.status()
             pct = status["currentLayer"] / status["totalLayers"]
-            progress.update(task, advance=status["currentLayer"]-previous_layer, completed=status["currentLayer"])
+            progress.update(task, advance=status["currentLayer"] - previous_layer, completed=status["currentLayer"])
             if pct >= 1.0:
                 break
             previous_layer = status["currentLayer"]
@@ -177,11 +209,7 @@ async def do_print(printer, filename):
         raise PrintError(msg)
 
 
-async def do_upload(
-    printer: SaturnPrinter,
-    filename: Path,
-    start_printing=False
-    ):
+async def do_upload(printer: SaturnPrinter, filename: Path, start_printing=False):
     if not Path(filename).exists():
         msg = f"{filename} does not exist"
         logger.error(msg)
@@ -216,7 +244,10 @@ async def do_upload(
     await upload_task
 
 
-def get_printers(printer: Optional[str] = None, broadcast: str = "<broadcast>",):
+def get_printers(
+    printer: Optional[str] = None,
+    broadcast: str = "<broadcast>",
+):
     if printer:
         printer = SaturnPrinter().find_printer(addr=printer)
         if printer is None:
@@ -232,12 +263,13 @@ def get_printers(printer: Optional[str] = None, broadcast: str = "<broadcast>",)
 @app.command(help="Discover and display status of all printers")
 def status(
     printer: Annotated[Optional[str], typer.Argument(help="ID of printer to target")] = None,
-    broadcast: Annotated[str, typer.Option(help="Explicit broadcast IP address")] = "<broadcast>",
-    status_full: Annotated[bool, typer.Option(help="Discover and display full status of all printers")] = False,
-    debug: Annotated[
-        bool,
-        typer.Option("--debug")
+    broadcast: Annotated[str, typer.Option("--broadcast", help="Explicit broadcast IP address")] = "<broadcast>",
+    status_full: Annotated[
+        bool, typer.Option("--full", help="Discover and display full status of all printers")
     ] = False,
+    live_update: Annotated[bool, typer.Option("--live", help="Update the status table in read time.")] = False,
+    update_interval: Annotated[int, typer.Option("--interval", help="Live update interval, in seconds.")] = 1,
+    debug: Annotated[bool, typer.Option("--debug")] = False,
 ):
     if debug:
         init_logger(3)
@@ -245,34 +277,37 @@ def status(
         printers = get_printers(printer=printer)
     else:
         printers = get_printers(broadcast=broadcast)
-    if status_full:
-        do_status_full(printers)
+    if live_update:
+        console = Console()
+        with Live(live_status(printers), console=console, refresh_per_second=4, transient=False, screen=False) as live:
+            while True:
+                time.sleep(update_interval)
+                live.update(do_status(printers))
+    elif status_full:
+        console.print(do_status_full(printers))
     else:
-        do_status(printers)
+        console.print(do_status(printers))
 
 
 @app.command(help="Continuously update the status of the selected printer")
 def watch(
-    printer_addr: Annotated[Optional[str], typer.Option("--printer",help="ID of printer to target")] = None,
+    printer_addr: Annotated[Optional[str], typer.Option("--printer", help="ID of printer to target")] = None,
     interval: Annotated[int, typer.Option("--interval", help="Status update interval (seconds)")] = 5,
-    debug: Annotated[
-        bool,
-        typer.Option("--debug")
-    ] = False,
+    debug: Annotated[bool, typer.Option("--debug")] = False,
 ):
     if debug:
         init_logger(3)
     do_watch(printer_addr, interval=interval)
 
+
 @app.command(help="Upload a file to the printer")
 def upload(
     filename: Annotated[Path, typer.Argument(help="File to upload")],
     printer_addr: Annotated[str, typer.Argument(help="ID of printer to target")],
-    start_printing: Annotated[bool, typer.Option("--start-printing", help="Start printing after upload is complete")] = True,
-    debug: Annotated[
-        bool,
-        typer.Option("--debug")
-    ] = False,
+    start_printing: Annotated[
+        bool, typer.Option("--start-printing", help="Start printing after upload is complete")
+    ] = True,
+    debug: Annotated[bool, typer.Option("--debug")] = False,
 ):
     if debug:
         init_logger(3)
@@ -285,14 +320,12 @@ def upload(
     else:
         asyncio.run(do_upload(printer, filename, start_printing=start_printing))
 
+
 @app.command(help="Start printing a file already present on the printer")
 def print_file(
     filename: Annotated[str, typer.Argument(help="File to print")],
     printer_addr: Annotated[str, typer.Argument(help="ID of printer to target")],
-    debug: Annotated[
-        bool,
-        typer.Option("--debug")
-    ] = False,
+    debug: Annotated[bool, typer.Option("--debug")] = False,
 ):
     if debug:
         init_logger(3)
@@ -305,27 +338,24 @@ def print_file(
     else:
         asyncio.run(do_print(printer, filename))
 
+
 @app.command(help="Connect printer to particular MQTT server")
 def connect_mqtt(
     address: Annotated[str, typer.Argument(help='MQTT host and port, e.g. "192.168.1.33:1883" or "mqtt.local:1883"')],
-    printer: Annotated[Optional[str], typer.Option("--printer",help="ID of printer to target")] = None,
+    printer: Annotated[Optional[str], typer.Option("--printer", help="ID of printer to target")] = None,
     broadcast: Annotated[Optional[str], typer.Option("--broadcast", help="Explicit broadcast IP address")] = None,
-    debug: Annotated[
-        bool,
-        typer.Option("--debug")
-    ] = False,
+    debug: Annotated[bool, typer.Option("--debug")] = False,
 ):
     if debug:
         init_logger(3)
     printers = get_printers(printer, broadcast)
 
     mqtt_host, mqtt_port = address.split(":")
-    try:
+    with contextlib.suppress(socket.gaierror):
         mqtt_host = socket.gethostbyname(mqtt_host)
-    except socket.gaierror:
-        pass
     for p in printers:
         p.connect_mqtt(mqtt_host, mqtt_port)
+
 
 if __name__ == "main":
     app()